refactor(web): extract WebFileInputSession to src/web_file_input_session.dart
