+++ packages/file_picker_web/pubspec.yaml
@@ -25,6 +25,7 @@ dependencies:
     sdk: flutter
   file_picker_platform_interface: ^1.0.0
   cross_file: ^0.3.5+2
+  meta: ^1.16.0
   path: ^1.9.0
   web: ^1.1.1
